backend/website-pf/pyproject.toml
[tool.poetry]
name = "website-react-pf-backend"
version = "1.0.0"
description = "website-react-pf-backend"
authors = [
    "Preston Frazier <preston@prestonfrazier.net>"
]
packages = [
   { include = "src" }
]

[tool.poetry.dependencies]
python = "~3.13"
boto3 = "1.42.45"
werkzeug = "3.1.5"
flask = "3.1.2"
flask-cors = "6.0.2"
serverless-wsgi = "3.1.0"
jsonpickle = "4.0.1"
mysql-connector-python = "9.1.0"
"ruamel.yaml" = "0.18.6"
pytz = "2025.2"
python-json-logger = "3.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.4"
pytest-cov = "6.0.0"
coverage = {extras = ["toml"], version = "7.6.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v --cov --cov-report term --cov-report xml"
log_cli = true
log_cli_level = "WARNING"
testpaths = [
    "tests/*"
]

[tool.coverage.run]
branch = true
include = [
    "src/*"
]
omit = [
    "*/tests/*",
    "*__init.__*"
]

[tool.coverage.xml]
output = "coverage.xml"
